+++ src/main/java/com/gmeister/temp/pkcmmsrando/rando/Randomiser.java
@@ -1,0 +1,203 @@
+package com.gmeister.temp.pkcmmsrando.rando;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Random;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+import com.gmeister.temp.pkcmmsrando.map.data.Block;
+import com.gmeister.temp.pkcmmsrando.map.data.BlockSet;
+import com.gmeister.temp.pkcmmsrando.map.data.Map;
+
+public class Randomiser
+{
+	
+	private Random random;
+	
+	/**
+	 * Initialises a Randomiser with a new Random.
+	 */
+	public Randomiser()
+	{
+		this.random = new Random();
+	}
+	
+	/**
+	 * Initialises a Randomiser with the provided Random.
+	 * @param random the Random to initialise with.
+	 */
+	public Randomiser(Random random)
+	{
+		this.random = random;
+	}
+	
+	/**
+	 * Shuffles blocks to other blocks with the same collision.
+	 * <br><br>
+	 * Replaces every block with a block that has the same collision values.
+	 * This creates maps that look like an absolute mess but play exactly like the original maps.
+	 * <br><br>
+	 * This randomiser is race safe.
+	 * @param blockSet the BlockSet from which new Blocks will be drawn
+	 * @param blocks the Blocks to replace
+	 * @return a new array of random blocks
+	 */
+	public Block[] randomiseBlocksByCollision(BlockSet blockSet, Block[] blocks)
+	{
+		Random random = new Random(this.random.nextLong());
+		Block[] randomisedBlocks = new Block[blocks.length];
+		
+		for (int i = 0; i < blocks.length; i++) if (blockSet.getBlocks().indexOf(blocks[i]) != 0)
+		{
+			ArrayList<Block> blockGroup = null;
+			for (ArrayList<Block> group : blockSet.getCollGroups()) if (group.contains(blocks[i]))
+			{
+				blockGroup = group;
+				break;
+			}
+			
+			int index = random.nextInt(blockGroup.size());
+			//inelegant solution, but works
+			while (blockSet.getBlocks().indexOf(blockGroup.get(index)) == 0) index = random.nextInt(blockGroup.size());
+			randomisedBlocks[i] = blockGroup.get(index);
+			if (randomisedBlocks[i] == null || blockSet.getBlocks().indexOf(randomisedBlocks[i]) == 0) throw new IllegalStateException();
+		}
+		
+		return randomisedBlocks;
+	}
+	
+	/**
+	 * Shuffles selected lines in a script.
+	 * @param script the script
+	 * @param toShuffle an array of the same length of the script, denoting the lines to shuffle
+	 * @return a shuffled script
+	 */
+	public ArrayList<String> shuffleScriptLines(ArrayList<String> script, boolean[] toShuffle)
+	{
+		Random random = new Random(this.random.nextLong());
+		ArrayList<String> shuffledScript = new ArrayList<>();
+		
+		for (int i = 0; i < toShuffle.length && i < script.size(); i++) if (toShuffle[i]) shuffledScript.add(script.get(i));
+		Collections.shuffle(shuffledScript, random);
+		for (int i = 0; i < script.size(); i++)
+		{
+			if (i >= toShuffle.length) shuffledScript.add(script.get(i));
+			else if (!toShuffle[i]) shuffledScript.add(i, script.get(i));
+		}
+		
+		return shuffledScript;
+	}
+	
+	/**
+	 * Shuffles selected lines in a script.
+	 * @param script the script
+	 * @param toShuffle a list of the same length of the script, denoting the lines to shuffle
+	 * @return a shuffled script
+	 */
+	public ArrayList<String> shuffleScriptLines(ArrayList<String> script, List<Boolean> toShuffle)
+	{
+		Random random = new Random(this.random.nextLong());
+		ArrayList<String> shuffledScript = new ArrayList<>();
+		
+		for (int i = 0; i < toShuffle.size() && i < script.size(); i++) if (toShuffle.get(i)) shuffledScript.add(script.get(i));
+		Collections.shuffle(shuffledScript, random);
+		for (int i = 0; i < script.size(); i++)
+		{
+			if (i >= toShuffle.size()) shuffledScript.add(script.get(i));
+			else if (!toShuffle.get(i)) shuffledScript.add(i, script.get(i));
+		}
+		
+		return shuffledScript;
+	}
+	
+	/**
+	 * Shuffles the pointers of all music tracks, such that every incidence of one track is replaced with a different track.
+	 * <br><br>
+	 * This randomiser is safe for races.
+	 * @param script the music script to shuffle
+	 * @return a shuffled script
+	 */
+	public ArrayList<String> shuffleMusicPointers(ArrayList<String> script)
+	{
+		boolean[] toShuffle = new boolean[script.size()];
+		for (int i = 0; i < script.size(); i++) toShuffle[i] = script.get(i).startsWith("\tdba") && !script.get(i).contains("Music_Nothing");
+		return this.shuffleScriptLines(script, toShuffle);
+	}
+	
+	/**
+	 * Shuffles the pointers of all sound effects, such that every incidence of one sound effect is replaced with a different sound effect.
+	 * <br><br>
+	 * This randomiser is safe for races, but having quick sound-effects replaced with long ones can greatly slow down pace.
+	 * @param script the SFX script to shuffle
+	 * @return a shuffled version of the provided script
+	 */
+	public ArrayList<String> shuffleSFXPointers(ArrayList<String> script)
+	{
+		boolean[] toShuffle = new boolean[script.size()];
+		for (int i = 0; i < script.size(); i++) toShuffle[i] = script.get(i).startsWith("\tdba");
+		return this.shuffleScriptLines(script, toShuffle);
+	}
+	
+	/**
+	 * Shuffles the pointers to all overworld sprites, such that all occurrences of each sprite are replaced with a different sprite.
+	 * <br><br>
+	 * This randomiser is currently untested.
+	 * @param script the overworld sprite script to shuffle
+	 * @return a shuffled version of the provided script
+	 */
+	public ArrayList<String> shuffleOverworldSpritePointers(ArrayList<String> script)
+	{
+		Pattern constPattern = Pattern.compile("^\\toverworld_sprite ");
+		List<Boolean> toShuffle = script.stream().map(s -> constPattern.matcher(s).find()).collect(Collectors.toList());
+		return this.shuffleScriptLines(script, toShuffle);
+	}
+	
+	/**
+	 * Move all trainers to a random position on the map.
+	 * <br><br>
+	 * This edits the map script in place and requires it to be written out.
+	 * <br><br>
+	 * This randomiser is not safe for races as trainers may block required paths.
+	 * @param map the map to shuffle trainer locations on
+	 */
+	public void randomiseTrainerLocation(Map map)
+	{
+		Random random = new Random(this.random.nextLong());
+		Pattern commentsPattern = Pattern.compile("\\s*;.*");
+		Pattern mapConstPattern = Pattern.compile("\\tmap_const\\s+");
+		Pattern commaWhitespacePattern = Pattern.compile("\\s*,\\s*");
+		Pattern objectEventPattern = Pattern.compile("\\tobject_event\\s+");
+		Pattern numberPattern = Pattern.compile("\\d+");
+		
+		for (int i = 0; i < map.getScript().size(); i++)
+		{
+			String line = map.getScript().get(i);
+			if (objectEventPattern.matcher(line).find())
+			{
+				String argsLine = commentsPattern.matcher(line).replaceFirst("");
+				argsLine = mapConstPattern.matcher(line).replaceFirst("");
+				String[] args = commaWhitespacePattern.split(argsLine);
+				
+				if (args[9].equals("OBJECTTYPE_TRAINER"))
+				{
+					int newX = random.nextInt(map.getXCapacity());
+					int newY = random.nextInt(map.getYCapacity());
+					
+					StringBuffer buffer = new StringBuffer();
+					Matcher numbers = numberPattern.matcher(line);
+					numbers.find();
+					numbers.appendReplacement(buffer, String.valueOf(newX));
+					numbers.find();
+					numbers.appendReplacement(buffer, String.valueOf(newY));
+					numbers.appendTail(buffer);
+					
+					map.getScript().set(i, buffer.toString());
+				}
+			}
+		}
+	}
+	
+}
